test constants package refactored
